feat(auth): refactor authentication forms and enhance session management
